requirements.txt
openai>=1.42.0
openai-whisper>=20231117
tiktoken>=0.7.0
typer[all]>=0.12.3
pydantic>=2.9.0
python-dotenv>=1.0.1
rich>=13.7.1
requests>=2.32.3
sounddevice>=0.4.9
soundfile>=0.13.1
PySide6>=6.7.2
pyinstaller>=6.0.0